+++ src/Synapse/Synapse/ChangeLog.md
@@ -23,6 +23,7 @@
 	- Added `Get-AzSynapseManagedPrivateEndpoint` cmdlet
 	- Added `Remove-AzSynapseManagedPrivateEndpoint` cmdlet
 * Fixed the blank page issue of pause setting and scale setting for Apache Spark pool through management API
+* Updated `Set-AzSynapseSqlActiveDirectoryAdministrator` to support for setting SQL Admin by `DisplayName` or by `ObjectId`
 
 ## Version 0.14.0
 * Added parameter `-ManagedResourceGroupName` for the `New-AzSynapseWorkspace` cmdlet